refactor: create a manager for user input validation
create a manager for user input validation

MODULE-1-PART-3-gallery
