+++ utilities/validation/emailValidation.js
@@ -1,15 +1,3 @@
-export function checkEmail(formElement, emailErrorElement) {
-    const email = formElement.email.value;
-
-    const validatedEmail = validateEmail(email);
-
-    if (!validatedEmail.isValid) {
-        emailErrorElement.innerHTML = 'Email is not valid!'
-    } else {
-        emailErrorElement.innerHTML = '';
-    }
-}
-
 export function validateEmail(email) {
     let userEmail = {
         isValid: false
