+++ utilities/validation/passwordValidation.js
@@ -1,15 +1,3 @@
-export function checkPassword(formElement, passwordErrorElement) {
-        const password = formElement.password.value;
-
-        const validatedPassword = validatePassword(password);
-
-        if (!validatedPassword.isValid) {
-            passwordErrorElement.innerHTML = validatedPassword.error;
-        } else {
-            passwordErrorElement.innerHTML = '';
-        }
-}
-
 export function validatePassword(p) {
     let error = '';
